+++ db/migrate/20210819171001_remove_notes_index.rb
@@ -1,0 +1,5 @@
+class RemoveNotesIndex < ActiveRecord::Migration[6.0]
+  def change
+    remove_index :notes, [:claim_id, :created_at]
+  end
+end